requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.5
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
psycopg2==2.8.5
pycodestyle==2.5.0
pytz==2020.1
sqlparse==0.3.1
